Drive Code + Constants
This is the first real commit for the year. It contains the Drive code
for this year's robot. A possible change we might have, is going from
a two stick drive to a one stick drive. This is tbd.

I also added all the button constants for future usage.

Beyond this point, it will mostly be the rest of you working on the
code.