Init Climb
I put some basic code in for this project. Obviously, we will have to
write a lot more than just this. Hopefully, this is a simple enough
launching pad so you know how to structure it.

Note, I put the climbing command in its own separate folder. I'm
predicting that it's going to take more than one command to accomplish
all the tasks for the climbing mechanism. Be sure to place all the
commands for the climb in there.