Cleaned + Testing Elevator
Cleaned code from the last commit. You guys need to document your
code as shown within this commit. Most importantly however, you
guys wrote a lot of unnecessary code for boolean stuff, and it
just makes the code hard to read.

If you look at the diffs, you'll actually see you had a dead
method, meaning that you had a method which did the same exact job
as the method referencing it.

I updated the current detection thing with a dummy value for the
current.