No More Gradle Errors
Due to Gradle files being created differently according to operating systems, and the fact that we have 3 different operating systems in the mix on the team right now, I'm deleting the gradle files from github. 

Laura (aka me because she doesn't know what a gitignore is) already altered the gitignore so the build files will no longer be picked up across the robots. Her edit also included the build.gradle file, which is a bit more drastic than needed, but it's not the worst to include. I believe another RoboRIO update is needed to support 2022.3.1 and the build.gradle file will be altered.