MASSIVE COMMIT READ BODY
Updated the Elevator and the Shooter/Susan

Elevator:
No longer automatic stopping based off current. Realized there was
no real use for it. The ball is gonna stop regardless at the
deadzone, so there's no reason to have this.

Shooter:
Pulled up the shooter code from last year, and have slightly
adapted it for this year. The linear speed method is the main one
and will let us do all of our calculations. We'll do some stuff
for it to work with enough time.

Lazy Susan:
Part of the shooter, but the basic aligning code is there. It's
not hooked up to a button for now though.